better code organization for pcp
